Removed aquifer.cpp and renamed aquifer to AquiferCT as it seems to be only relevant for CarterTracy
